requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
notion-client==2.2.1
requests
pydantic
